requirements.txt
# Production-ready Python dependencies
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.28.0
pydantic==2.6.3
pydantic-settings==2.2.1

# Celery & Queue
celery==5.3.6
redis==5.0.3

# Database
pymongo==4.6.2
motor==3.4.0

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.1
requests==2.31.0

# AI/ML
numpy==1.26.4
langchain==0.1.12
langchain-openai==0.0.8
pinecone==8.0.0
PyPDF2==3.0.1
pydantic-settings==2.2.1
cohere==5.0.0
